+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## [Unreleased]
 
-* Added `decode_hash` and `original` decoders
+* Added `decode_hash`, `regexp` and `original` decoders
 
 ## [0.1.0] - 2024-05-08